task-13_

+++ 0x00-MySQL_Advanced/101-average_weighted_score.sql
@@ -6,7 +6,7 @@ DELIMITER $$
 
 CREATE PROCEDURE ComputeAverageWeightedScoreForUsers ()
 BEGIN
-    ALTER TABLE users ADD total_weighted_score FLOAT NOT NULL;
+    ALTER TABLE users ADD total_weighted_score INT NOT NULL;
     ALTER TABLE users ADD total_weight INT NOT NULL;
 
     UPDATE users
